+++ integrations/azure_ai_search/tests/conftest.py
@@ -28,6 +28,7 @@ def document_store(request):
     """
     index_name = f"haystack_test_{uuid.uuid4().hex}"
     metadata_fields = getattr(request, "param", {}).get("metadata_fields", None)
+    include_search_metadata = getattr(request, "param", {}).get("include_search_metadata", False)
 
     azure_endpoint = os.environ["AZURE_AI_SEARCH_ENDPOINT"]
     api_key = os.environ["AZURE_AI_SEARCH_API_KEY"]
@@ -37,12 +38,11 @@ def document_store(request):
         client.delete_index(index_name)
 
     store = AzureAISearchDocumentStore(
-        api_key=api_key,
-        azure_endpoint=azure_endpoint,
         index_name=index_name,
         create_index=True,
         embedding_dimension=768,
         metadata_fields=metadata_fields,
+        include_search_metadata=include_search_metadata,
     )
 
     # Override some methods to wait for the documents to be available